chore(forknet): update example
